error handling initial

+++ PAF.DAS.WebAPI/Controllers/PapersController.cs
@@ -69,36 +69,43 @@ namespace PAF.DAS.WebAPI.Controllers
         public IActionResult Post([FromBody]PaperArchiveModel value)
         { 
             var paper = _mapper.Map<PaperArchiveModel, Paper>(value);
-            var result = _paperService.Add(paper);
-            if (result != null)
+            if (!ValidateTitle(paper.Title))
             {
-                string tempPath = Path.Combine(Path.GetTempPath(), value.FileName.ToString());
-                string uploadPath = Path.Combine(_hostingEnvironment.WebRootPath, "uploads");
-                string officialPath = Path.Combine(uploadPath, value.FileName.ToString());
-                _paperArchiveService.Add(new PaperArchieve
+                var result = _paperService.Add(paper);
+                if (result != null)
                 {
-                    FileName = value.FileName.ToString(),
-                    Location = officialPath,
-                    PaperId = result.Id
-                });
-
-                FileInfo file = new FileInfo(tempPath);
-                DirectoryInfo uploadDirectory = new DirectoryInfo(uploadPath);
-                if(file.Exists)
-                {
-                    if (!uploadDirectory.Exists)
+                    string tempPath = Path.Combine(Path.GetTempPath(), value.FileName.ToString());
+                    string uploadPath = Path.Combine(_hostingEnvironment.WebRootPath, "uploads");
+                    string officialPath = Path.Combine(uploadPath, value.FileName.ToString());
+                    _paperArchiveService.Add(new PaperArchieve
                     {
-                        uploadDirectory.Create();
-                    }
-                    file.CopyTo(officialPath);
-                    file.Delete();
-                }
+                        FileName = value.FileName.ToString(),
+                        Location = officialPath,
+                        PaperId = result.Id
+                    });
 
-                return Ok(result);
+                    FileInfo file = new FileInfo(tempPath);
+                    DirectoryInfo uploadDirectory = new DirectoryInfo(uploadPath);
+                    if (file.Exists)
+                    {
+                        if (!uploadDirectory.Exists)
+                        {
+                            uploadDirectory.Create();
+                        }
+                        file.CopyTo(officialPath);
+                        file.Delete();
+                    }
+
+                    return Ok(result);
+                }
+                else
+                {
+                    return BadRequest("Adding paper that are already exist is not permitted.");
+                }
             }
             else
             {
-                return StatusCode(409);
+                return BadRequest("Adding paper that are already exist is not permitted.");
             }
         }
 
@@ -116,6 +123,10 @@ namespace PAF.DAS.WebAPI.Controllers
             {
                 return StatusCode(404);
             }
-        }       
+        }
+        private bool ValidateTitle(string title)
+        {
+            return _paperService.GetAll().Any(p => p.Title.ToLower() == title.ToLower());
+        }
     }
 }
